+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2024-11-10
+### Added
+- New test
+
+### Changed
+- Improved error context capture (preparation for v0.3.0)
+- Source code verification removed (seems unnecessary)
+
 ## [0.2.3] - 2024-11-05
 ### Added
 - Saving global and local variables to context